Ambalan Ventures/Norpro/norpro-fe_be-django_html-master/norpro-fe_be-django_html-master/requirements.txt
asgiref==3.7.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==44.0.2
Django==4.2.7
django-appconf==1.0.6
django-ckeditor==6.7.0
django-cors-headers==4.6.0
django-embed-video==1.4.10
django-js-asset==2.1.0
django-select2==8.2.3
django-taggit==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
firebase-admin==6.6.0
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.7
msgpack==1.1.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
qrcode==7.4.2
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.2
vobject==0.9.6.1
